feature(today-avarage-component): create constants files
