+++ challenge-catar Watch App/Constants/ColorConstant.swift
@@ -1,0 +1,14 @@
+//
+//  ColorConstant.swift
+//  challenge-catar Watch App
+//
+//  Created by alexdamascena on 27/11/22.
+//
+
+import Foundation
+
+struct ColorConstant {
+    static let BLUE = "blue"
+    static let LIGHT_GRAY = "light_gray"
+    static let PURPLE = "purple"
+}
