+++ challenge-catar Watch App/Views/Components/CardInformation/TitleCardInformation.swift
@@ -1,0 +1,13 @@
+//
+//  TitleCardInformation.swift
+//  challenge-catar Watch App
+//
+//  Created by alexdamascena on 27/11/22.
+//
+
+import Foundation
+
+enum TitleCardInformation: String {
+    case today = "Hoje"
+    case avarage = "Média Semanal"
+}
